requirements.txt
azure-identity>=1.15.0
requests>=2.32.3
python-dotenv>=1.0.0
